+++ api/code/firewall.go
@@ -1440,6 +1440,10 @@ func populateSets() {
 		}
 		if iface.Type == "Uplink" && iface.Enabled == true {
 			addUplinkEntry(iface.Name, iface.Subtype, false)
+			//Set interfaces up
+			if iface.Subtype != "pppup" && isValidIface(iface.Name) {
+				exec.Command("ip", "link", "set", iface.Name, "up").Run()
+			}
 		}
 	}