+++ frontend/src/views/AuthSettings.js
@@ -5,6 +5,7 @@ import { View } from '@gluestack-ui/themed'
 import WebAuthn from 'components/Auth/WebAuthn'
 import AuthTokenList from 'components/Auth/AuthTokenList'
 import OTPSettings from 'components/Auth/OTPSettings'
+import SSHKeys from 'components/Auth/SSHKeys'
 
 export default class AuthSettings extends Component {
   constructor(props) {
@@ -17,6 +18,7 @@ export default class AuthSettings extends Component {
         <OTPSettings />
         {/*<WebAuthn />*/}
         <AuthTokenList />
+        <SSHKeys />
       </View>
     )
   }